// app/Core/Enums/StatType.ts
export enum StatType {
  NONE = 'NONE',
  CRIT_DMG = 'CRIT_DMG',
  CRIT_RATE = 'CRIT_RATE',
  ATTACK = 'ATTACK',
  ATTACK_PERCENTAGE = 'ATTACK_PERCENTAGE',
  HP = 'HP',
  HP_PERCENTAGE = 'HP_PERCENTAGE',
  HEALING_BONUS = 'HEALING_BONUS',
  DEF = 'DEF',
  DEF_PERCENTAGE = 'DEF_PERCENTAGE',
  ENERGY_REGENERATION = 'ENERGY_REGENERATION',
  BASIC_ATTACK_DMG_AMPLIFICATION = 'BASIC_ATTACK_DMG_AMPLIFICATION',
  HEAVY_ATTACK_DMG_AMPLIFICATION = 'HEAVY_ATTACK_DMG_AMPLIFICATION',
  RESONANCE_LIBERATION_DMG_AMPLIFICATION = 'RESONANCE_LIBERATION_DMG_AMPLIFICATION',
  RESONANCE_SKILL_DMG_AMPLIFICATION = 'RESONANCE_SKILL_DMG_AMPLIFICATION',
  ELECTRO_DMG_BONUS = 'ELECTRO_DMG_BONUS',
  FUSION_DMG_BONUS = 'FUSION_DMG_BONUS',
  GLACIO_DMG_BONUS = 'GLACIO_DMG_BONUS',
  HAVOC_DMG_BONUS = 'HAVOC_DMG_BONUS',
  SPECTRO_DMG_BONUS = 'SPECTRO_DMG_BONUS',
}